Update ProductPageVideos.html
Changed a lot -- mainly moved the "video detector" function out of document ready and changed the click() function to an on() function.  

The change in placement places the detector function out of the document ready so that it doesn't wait for the document to render, and just goes ahead and finds the video demo trigger and continues to inject the HTML for the button Play/Stop trigger.

The change from click() to the delegated on() function with a 'click' event allows the jQuery to edit the HTML which is dynamically added with jQuery.  Now, the button is added, the document finishes loading, then finally the Play/Stop functionality of the jQuery loads and can now edit the dynamically created elements by the jQuery button detector.