Improved behavior of #videoDemoThumb removal when no video present
Conflicting with MagicScroll, as MagicScroll makes a clone of the list
of items before my JS has a chance to detect no video and to remove the
useless video button.
— Deleting from the live list become impossible as MS just keeps
re-creating the items as soon as the scroll is moved.
— Only solution at the moment is to wait for $(window).load() before
proceeding to MagicScroll.stop(), removing the empty button, and then
MagiScroll.start();
— if I try to stop the MS before window load, an error is thrown that
“can’t stop undefined”…